fix: add cloud storage trait
