+++ Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+async-trait = "0.1.86"
 async-walkdir = "1"
 async-recursion = "1.0.0"
 aws-config = "1"
