+++ src/domain/mod.rs
@@ -1,0 +1,3 @@
+mod cloud_storage;
+
+pub use cloud_storage::*;
