+++ src/lib.rs
@@ -1,3 +1,5 @@
+pub mod cloud_storage;
+pub mod domain;
 pub mod error;
 pub mod sync_tool;
 pub mod utils;
