+++ src/main.rs
@@ -2,7 +2,9 @@ use std::{env, time::Instant};
 
 use color_eyre::Result;
 
-use sync_tool::sync_tool::{dif, show, sync, upload, Mode};
+use sync_tool::cloud_storage::aws::AwsStorage;
+use sync_tool::domain::CloudStorage;
+use sync_tool::sync_tool::Mode;
 use sync_tool::utils::aws::get_aws_client;
 use sync_tool::utils::constants::*;
 
@@ -15,6 +17,9 @@ async fn main() -> Result<()> {
     let _ = args.next();
     let mode = args.next().unwrap_or(DEFAULT_MODE.to_string());
     let config = CONFIG.to_owned();
+    let client = get_aws_client(&config.region).await;
+    let aws_storage = AwsStorage::new(client);
+
     if let Some(mode) = Mode::new(&mode) {
         let source = config.source.to_string();
         let target = format!("s3://{}/{}", &config.bucket, &config.target);
@@ -24,12 +29,11 @@ async fn main() -> Result<()> {
             &source,
             &target
         );
-        let client = get_aws_client(&config.region).await;
         match mode {
-            Mode::Dif => dif(client, config).await?,
-            Mode::Upload => upload(client, config).await?,
-            Mode::Sync => sync(client, config).await?,
-            Mode::Show => show(client, config).await?,
+            Mode::Dif => aws_storage.dif(&config).await?,
+            Mode::Upload => aws_storage.upload(&config).await?,
+            Mode::Sync => aws_storage.sync(&config).await?,
+            Mode::Show => aws_storage.show(&config).await?,
         }
         println!(
             "sync-tool finished with mode: {} for source: {} target: {} elapsed: {:.2?}",
