+++ src/sync_tool.rs
@@ -1,19 +1,3 @@
-use std::{collections::HashMap, ffi::OsStr, path::Path};
-
-use crate::utils::{
-    aws::{list_keys_stream, upload_object, upload_object_multipart},
-    config::AppConfig,
-    constants::FILES_TO_IGNORE,
-    tools::sanitize_file_path,
-};
-use crate::Result;
-
-use async_recursion::async_recursion;
-use async_walkdir::{Filtering, WalkDir};
-use aws_sdk_s3::Client;
-use tokio::{fs, task::JoinSet};
-use tokio_stream::StreamExt;
-
 /// Mode has 4 options:
 /// dif - calculate and show dif in file name and size
 /// upload - simple upload files without checking target file names and sizes
@@ -53,262 +37,4 @@ impl Mode {
             }
         }
     }
-}
-
-/// Calculate dif and print
-pub async fn dif(client: Client, config: AppConfig) -> Result<()> {
-    let source_task = tokio::spawn(files_walker(config.source.clone()));
-    let target_task = tokio::spawn(list_keys_stream(
-        client.clone(),
-        config.bucket.clone(),
-        config.target.clone(),
-    ));
-    let (source, target) = (source_task.await??, target_task.await??);
-
-    let dif = dif_calc(&config, &source, &target);
-    match dif {
-        Some(dif) => println!("dif found: {:?}", dif),
-        None => println!("no dif found"),
-    };
-
-    Ok(())
-}
-
-fn dif_calc(
-    config: &AppConfig,
-    source: &HashMap<String, i64>,
-    target: &HashMap<String, i64>,
-) -> Option<HashMap<String, i64>> {
-    let mut dif: HashMap<String, i64> = HashMap::new();
-    for (k, v) in source.iter() {
-        let source_f_name = k
-            .strip_prefix(&config.source)
-            .unwrap_or("no_file_name")
-            .to_string();
-        let source_f_size = *v;
-        let target_f_name = format!("{}{}", &config.target, &source_f_name);
-        match target.get(&target_f_name) {
-            Some(target_f_size) => {
-                if source_f_size != *target_f_size {
-                    dif.insert(source_f_name.clone(), source_f_size);
-                }
-            }
-            None => {
-                dif.insert(source_f_name.clone(), source_f_size);
-            }
-        }
-    }
-
-    if dif.is_empty() {
-        None
-    } else {
-        Some(dif)
-    }
-}
-
-async fn files_walker<P>(path: P) -> Result<HashMap<String, i64>>
-where
-    P: AsRef<Path> + std::marker::Send + std::marker::Sync + std::fmt::Debug,
-{
-    #[async_recursion]
-    async fn files_walker_inner<P>(path: P, files: &mut HashMap<String, i64>) -> Result<()>
-    where
-        P: AsRef<Path> + std::marker::Send + std::marker::Sync + std::fmt::Debug,
-    {
-        let mut entries = fs::read_dir(&path).await?;
-        while let Some(entry) = entries.next_entry().await? {
-            if entry.path().is_file() {
-                if FILES_TO_IGNORE.iter().any(|f| {
-                    entry
-                        .path()
-                        .file_name()
-                        .unwrap_or(OsStr::new("no_file_name"))
-                        .to_string_lossy()
-                        == *f
-                }) {
-                    continue;
-                }
-
-                let file_name = entry.path().to_string_lossy().to_string();
-                let file_name = sanitize_file_path(&file_name); // sanitize file_name for windows only onces here
-                let file_size = entry.metadata().await?.len() as i64;
-                files.insert(file_name, file_size);
-            } else if entry.path().is_dir()
-            // && !entry.file_name().to_string_lossy().starts_with(".DS_Store") // #TODO seperate files and folders to ignore?
-            {
-                files_walker_inner(entry.path(), files).await?;
-            }
-        }
-        Ok(())
-    }
-
-    let mut files: HashMap<String, i64> = HashMap::new();
-    files_walker_inner(path, &mut files).await?;
-    Ok(files)
-}
-
-/// Print files from source and target folders
-pub async fn show(client: Client, config: AppConfig) -> Result<()> {
-    let source_task = tokio::spawn(files_walker(config.source));
-    let target_task = tokio::spawn(list_keys_stream(client, config.bucket, config.target));
-    let (source, target) = (source_task.await??, target_task.await??);
-    println!("source: {:?}", source);
-    println!("target: {:?}", target);
-    Ok(())
-}
-
-/// Upload files from source to target without checking file name and size
-pub async fn upload(client: Client, config: AppConfig) -> Result<()> {
-    let mut entries = WalkDir::new(&config.source).filter(|entry| async move {
-        if let Some(true) = entry
-            .path()
-            .file_name()
-            .map(|file| FILES_TO_IGNORE.iter().any(|f| file.to_string_lossy() == *f))
-        {
-            return Filtering::IgnoreDir;
-        }
-        Filtering::Continue
-    });
-
-    let mut tasks = JoinSet::new();
-    let chunk_size = config.chunk_size * 1024 * 1024; // MB
-
-    while let Some(entry) = entries.next().await.transpose()? {
-        if let Ok(file) = entry.file_type().await {
-            if file.is_file() {
-                let file_name = entry.path().to_string_lossy().to_string();
-                let file_name = sanitize_file_path(&file_name); // sanitize file_name for windows only onces here
-                let file_size = entry.metadata().await?.len();
-                let file_name = Path::new(&file_name).strip_prefix(&config.source)?;
-                let file_name = file_name.to_string_lossy().to_string();
-                let key = format!("{}{}", &config.target, &file_name);
-                let f_name = format!("{}{}", &config.source, &file_name);
-
-                if file_size < chunk_size as u64 {
-                    tasks.spawn(upload_object(
-                        client.clone(),
-                        config.bucket.clone(),
-                        f_name,
-                        key,
-                        file_size,
-                    ))
-                } else {
-                    tasks.spawn(upload_object_multipart(
-                        client.clone(),
-                        config.bucket.clone(),
-                        f_name,
-                        key,
-                        file_size,
-                        chunk_size as u64,
-                        config.max_chunks as u64,
-                    ))
-                };
-
-                if tasks.len() == config.workers {
-                    if let Some(res) = tasks.join_next().await {
-                        match res {
-                            Ok(res) => {
-                                if let Err(e) = res {
-                                    println!("could not upload object: {}", e);
-                                }
-                            }
-                            Err(e) => println!("failed running task: {}", e),
-                        }
-                    }
-                }
-            }
-        }
-    }
-    while let Some(res) = tasks.join_next().await {
-        match res {
-            Ok(res) => {
-                if let Err(e) = res {
-                    println!("could not upload object: {}", e);
-                }
-            }
-            Err(e) => println!("could not run task: {}", e),
-        }
-    }
-    Ok(())
-}
-
-/// Sync files from source to target with checking file name and size
-pub async fn sync(client: Client, config: AppConfig) -> Result<()> {
-    let target =
-        list_keys_stream(client.clone(), config.bucket.clone(), config.target.clone()).await?;
-
-    let mut entries = WalkDir::new(&config.source).filter(|entry| async move {
-        if let Some(true) = entry
-            .path()
-            .file_name()
-            .map(|file| FILES_TO_IGNORE.iter().any(|f| file.to_string_lossy() == *f))
-        {
-            return Filtering::IgnoreDir;
-        }
-        Filtering::Continue
-    });
-
-    let mut tasks = JoinSet::new();
-    let chunk_size = config.chunk_size * 1024 * 1024; // MiB
-
-    while let Some(entry) = entries.next().await.transpose()? {
-        if let Ok(file) = entry.file_type().await {
-            if file.is_file() {
-                let file_name = entry.path().to_string_lossy().to_string();
-                let file_name = sanitize_file_path(&file_name); // sanitize file_name for windows only onces here
-                let source_file_size = entry.metadata().await?.len();
-                let source_file_name = Path::new(&file_name).strip_prefix(&config.source)?;
-                let source_file_name = source_file_name.to_string_lossy().to_string();
-                let target_file_name = format!("{}{}", &config.target, &source_file_name);
-                let target_file_size = target.get(&target_file_name).unwrap_or(&0);
-
-                if source_file_size != *target_file_size as u64 {
-                    // dif found
-                    if source_file_size < chunk_size as u64 {
-                        tasks.spawn(upload_object(
-                            client.clone(),
-                            config.bucket.clone(),
-                            file_name,
-                            target_file_name,
-                            source_file_size,
-                        ))
-                    } else {
-                        tasks.spawn(upload_object_multipart(
-                            client.clone(),
-                            config.bucket.clone(),
-                            file_name,
-                            target_file_name,
-                            source_file_size,
-                            chunk_size as u64,
-                            config.max_chunks as u64,
-                        ))
-                    };
-
-                    if tasks.len() == config.workers {
-                        if let Some(res) = tasks.join_next().await {
-                            match res {
-                                Ok(res) => {
-                                    if let Err(e) = res {
-                                        println!("could not upload object: {}", e);
-                                    }
-                                }
-                                Err(e) => println!("failed running task: {}", e),
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    while let Some(res) = tasks.join_next().await {
-        match res {
-            Ok(res) => {
-                if let Err(e) = res {
-                    println!("could not upload object: {}", e);
-                }
-            }
-            Err(e) => println!("could not run task: {}", e),
-        }
-    }
-    Ok(())
 }
