+++ sync-tool.json
@@ -1,9 +1,9 @@
 {
-    "bucket": "foo",
-    "region": "bar",
-    "source": "/path/to/source/data/",
-    "target": "path/to/target/data/",
-    "workers": 5,
+    "bucket": "dmg-data-dev",
+    "region": "eu-central-1",
+    "source": "/Users/szhukov/Documents/Docs/demo/considering-rust/",
+    "target": "szhukov/data/demo/considering-rust/",
+    "workers": 4,
     "chunk_size": 10,
     "max_chunks": 10000
 }
